ITF componentization
