+++ interface_board/CMakeLists.txt
@@ -5,5 +5,24 @@
 cmake_minimum_required(VERSION 3.5)
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-# include_directories(${CMAKE_CURRENT_LIST_DIR}/conf)
+# include_directories(${CMAKE_CURRENT_LIST_DIR}/main/conf)
+# Flags for everyone (C and C++)
+add_compile_options(
+    -Wno-missing-field-initializers
+    -Wno-unused-but-set-variable
+    -Wno-unused-variable
+    -Wno-unused-function
+    -Wno-extra
+)
+
+# Flags ONLY for C++
+add_compile_options(
+    $<$<COMPILE_LANGUAGE:CXX>:-Wno-deprecated-enum-enum-conversion>
+    # $<$<COMPILE_LANGUAGE:CXX>:-Wno-volatile>
+)
+
+# Flags ONLY for C
+# add_compile_options(
+#     $<$<COMPILE_LANGUAGE:C>:-Wno-pointer-sign>
+# )
 project(interface_board)
