+++ interface_board/main/mcpwm_capture_helpers.h
@@ -1,10 +1,11 @@
 #pragma once
 #include <stdio.h>
 #include "esp_log.h"
+#include "driver/gpio.h"
 #include "driver/mcpwm_prelude.h"
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
-#include "esp_timer.h"
+// #include "esp_timer.h"
 #include "driver/gptimer.h"
 
 #ifdef TAG
