+++ interface_board/main/version_parser.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <stdio.h>
 #include <string.h>
 #include "esp_app_desc.h"
